refactor : change directory structure
