+++ src/components/atoms/Stamp.jsx
