+++ src/components/features/homePage/HomePage.jsx
@@ -2,8 +2,8 @@ import Carousel from "../carousel/Carousel";
 import { useQuery } from "react-query";
 import { getHome } from "../../../apis/home";
 import HorizontalListSection from "../carousel/HorizontalListSection";
-import FestivalPoster from "../../molecules/FestivalPoster";
-import PlacePoster from "../../molecules/PlacePoster";
+import FestivalPoster from "../../molecules/cards/FestivalPoster";
+import PlacePoster from "../../molecules/cards/PlacePoster";
 import { utils } from "./utils";
 import SectionTitle from "../../atoms/SectionTitle";
 import HorizontalListSectionMediaQuery from "../carousel/HorizontalListSectionMediaQuery";
